First pass at hiding GPU offload/tracking manager impl.
Sketch out `GPUOffloadInterface` that is basically a pure virtual
virtual version of Celeritas' `SimpleOffload`. It abstracts the
member functions needed/thought to be needed for offloading to
GPU with:

- "Build" functions to be called in the corresponding UserActionInitialisation
  calls
- Syncronization/notification to Geant4's Run/Event begin/end
points.
- Factory function to return a concrete G4VTrackingManager for doing
the offload

An implementation of this interface for Celeritas is implemented,
following the setup shown in the Celeritas examples and its use
in the ATLAS TileCal test beam example.

The offload interface is accessed through a front end function,
that currently only sets up/returns the Celeritas implementation.
An AdePT implementation and setup is forthcoming.