Provide interfaces to choose offload backend at runtime
As outlined in Issue #5, the type of GPU offload implementation
(none, adept, or celeritas) is currently hardcoded at compile time.
For testing and use locally and by experiments, this should be
selectable at runtime, with additional configuration options passed
to the selected backend as needed.

Implement basic runtime selection of offloading backend via:

- An `enum class` representing the available backends, with an
  "unknown" entry to represent "unconfigured" state.
- A struct to hold the selected backend as a parameter.
- A user function that should be called at application start
  (strictly during Geant4's "PreInit" state) to setup the
  chosen backend.
- Private functions to construct required thread-local static
  instances of the selected backend, given the static global
  configuration.

A minimal demo of use is implemented in the smoke test application.
Whilst still hardcoded, this is now in a location that can be
configured in a variety of ways later on.